early new menu test

+++ src/game/q_shared.h
@@ -229,7 +229,7 @@ SYSTEM SPECIFIC
 */
 
 extern	int32_t curtime;		// time returned by last Sys_Milliseconds
-extern	int32_t curtime_ns;		// time returned by last Sys_Nanoseconds
+extern	int64_t curtime_ns;		// time returned by last Sys_Nanoseconds
 
 int32_t 	Sys_Milliseconds();
 int64_t		Sys_Nanoseconds(); // should be platform independent
